Added function kirbytext_placeholders()
kirbytext_placeholders() is a function for changing content in pages formerly generated by the kirbytext() function with placeholders.
Simply include any text in your Markdown text which you want to use as a placeholder and then call
	kirbytext_placeholders($text, array("Placeholder (any text)" => "Text you want to have in the output instead of placeholder", "Another one" => "Something"));
to change some vars dynamically, for example if you want to have dates included in your texts.

Maybe it can help somebody else.