Added *global* placeholders using YAML files
That's a major breakthrough!
You can now manage domain-specific global placeholders.

There's a new "replace"-folder in the "site"-folder, in which you can put .php files (written in YAML) with the global placeholders, you can even use PHP inside, as you can in a template (also with helper functions as url() and access to all variables).
That works like the config files, there is a global replace.php file, but you can also create a URL-specific version (replace.example.com.php) overwriting the placeholders in the replace.php (like if I were included after the replace.php).

The content has to be in YAML, like this:
<code>
replace this:
  with: replace with this
  usage: global
</code>
The first line is the String you want to replace, the second is the String to insert (only here you can use PHP, the other options wouldn't make any sense to have them dynamically), and as "usage" you can use "global", "kirbytext" and "demand".

"global" will be used on your whole source code (the final page).
"kirbytext" will be used in all kirbytext() calls, you don't have to give this placeholder to kirbytext, it just works.
"demand" will work in all kirbytext() calls like this:
<code>
kirbytext($text, array(
	"to replace",
	"another one"
));
</code>
It will automatically fetch the replacement from the global file (so don't use "to replace" => "anything", in this case "anything" would be used).
You can also define an alias in the YAML file like this:
<code>
this is a long text:
  with: anything
  usage: demand
  alias: replace
</code>
When you then call
<code>
kirbytext($text, array(
	"replace"
));
</code>
all "this is a long text" will be replaced with "anything".

That's it. Cool, eh? :)

PS: I know that I can't use GitHub Markdown in Commit messages, but anyway, <code> is just cool and better to read.