Now with template defining
You can now even just include some templates (for example if you only use the replacement in a specific template).

Just create a structure like this:
<code>
replace this:
  with: that
  usage: (all will work)
  templates:
    default
    blog
    about-us
</code>
That will only work with an array structure, anything else (strings) will be ignored.
All templates that either exist in the templates folder or just files with a specific name (like home.txt or about-us.txt) will be recognized. So you can also just rename files, automatically use the default template but define specific replacements.

I also fixed a bug in "demand" without existing one and nonexistent alias with that name. Now it won't echo errors (it checks the availability of the alias variable first).

I hope that is the last commit for the "placeholders" function...