- Added LESS compiler - Added better minifiers - Bug fixes
Now there's a new function called less() working like css() to let Kirby auto compile LESS files and cache them (so that only works with 'cache' set to true).

The new minifier 'HTML' is not used in the moment, we'll have to wait until the placeholders are finished and published so that I can re-fork the repo and add the support for this (it requires an addition to global placeholders).

There are also some bug fixes, I tested all the features, now they work (I hope.).