Added a "autoconfig" feature for plugins
With this commit, every plugin having a config.php with default values for that plugin in its folder (Addition in Pull Request #21) will be copied to the config dir (if writable) and autoloaded.

An example:
Plugin flickrbadge has the following file:
site/plugins/flickrbadge/config.php

It will be ignored as a plugin file but copied to
site/config/flickrbadge.php

where you can change the values as a user.
This will just happen one time (when the files does not exist in the config dir).
It's also possible to duplicate these files to a file like that:
site/config/flickrbadge.example.com.php

These will be loaded after the one without server name.