Added ability for plugins to have an assets folder
Now every plugin can have a folder named "assets" inside of the plugin folder with CSS-, JS-files and images and whatever, which is accessible over the internet (there is a new RewriteCond for this).

To make it a lot easier for plugins to get the right URL to an asset, it can now just call "getpluginasset($name)" to get a specific file URL, for example "getpluginasset('file.css')" will return "/site/plugins/ANYTHING/assets/file.css".
The great thing: The current plugin is auto detected, so there is no  need to pass the name manually. So plugins can even be installed to another folder (I will make a pull request soon) and it will still work.

Cool, em?