Multiple plugin folders
With this feature the user can add another (or multiple) new folders searched for plugins.

This is great, if the user has multiple sites powered by the same Kirby installation and wants both sites to have a specific plugin. Instead of installing (and updating…) the plugin in both site/plugins/ folders or even symlink it, you can now define all folders with plugins in the config.php of each site.

The given folder names can either be relative to the Kirby installation or absolute (to the partition of course :D).