Removed that strange design predecision for Kirbytext code blocks
The highlight() function must now return the whole code including <pre> and <code>, because it was just a thing some people might not want to have.

I need a specific class in the <pre> tag to make this work with the beautiful Prism.js, so it would be great, if I had the ability to do that.

Now, you can easily return whatever you want.

Will change that for the highlight plugin as well.